python_backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain
langchain>=0.1.20
langchain-openai>=0.1.3
langchain-community>=0.0.38
langchain-anthropic>=0.1.4
langchain-google-genai>=1.0.1
langchain-text-splitters>=0.0.1

# Embeddings & Vector Stores
openai>=1.12.0
chromadb>=0.4.22
faiss-cpu>=1.7.0
sentence-transformers>=2.3.1

# Document Processing
pypdf>=4.0.1
pymupdf>=1.23.8
beautifulsoup4>=4.12.3
requests>=2.31.0

# Text Processing
tiktoken>=0.5.2
rank-bm25>=0.2.2
kiwipiepy>=0.16.1

# Evaluation
ragas>=0.1.5
datasets>=2.16.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
aiofiles>=23.2.1
